Have functions return typed output
The LLMs seem to do well with stderr / stdout / exit_code
